Estimate the Nusselt number from the Reynolds number (based on Mitchell 1976)

Estimate the Nusselt number from the Reynolds number (based on Mitchell 1976)

Nu_from_Re(Re, taxa = "cylinder")

Arguments

Re
is the Reynolds Number (dimensionless)

taxa
Which class of organism, current choices: sphere, cylinder, frog, lizard_traverse_to_air_flow, lizard_parallel_to_air_flow, lizard_surface, lizard_elevated, flyinginsect, spider

Value

Nusselt number (dimensionless)

Details

This function allows you to estimate the Nusselt number from the Reynolds number for various taxa. Source: Mitchell JW. 1976. Heat transfer from spheres and other animal forms. Biophysical Journal 16: 561-569. See Table 1. Convective Heat Trasfer Relations for Animal Shapes.

See also

Examples

if (FALSE) {
Nu_from_Re(Re=5, taxa="cylinder")
}