Calculate basal (or resting) metabolism as a function of mass and body temperature.

estimate basal (or resting) metabolic rate (W) as a function of mass (g) and temperature (K). Based on empirical data and the metabolic theory of ecology (3/4 scaling exponent). Reference: Gillooly JF et al. 2001. Effects of size and temperature on metabolic rate. Science 293: 2248-2251.

Qmetabolism_from_mass_temp(m, T_b, taxa)

Arguments

m
Mass in grams.

T_b
body temperature in K

taxa
Which taxa. options: bird, mammal, reptile, amphibian, invertebrate.

Value

basal metabolim (W)

See also

Examples

Qmetabolism_from_mass_temp(m=100, T_b=303, "reptile")

#> [1] 0.06282736