Absorbed Thermal Radiation

The function estimates longwave (thermal) radiation (W) absorbed from the sky and the ground (Campbell and Norman 1998; Riddell et al. 2018) .

Qthermal_radiation_absorbed(
  T_a,
  T_g,
  epsilon_ground = 0.97,
  a_longwave = 0.965
)

Arguments

T_a

numeric air temperature (C).

T_g

numeric ground temperature (C).

epsilon_ground

numeric emissivity (proportion) for more soil types (Campbell and Norman 1998) .

a_longwave

numeric absorptance (proportion) of organism to longwave radiation (Bartlett and Gates 1967; Buckley 2008) .

Value

numeric thermal radiation absorbed (W).

References

Bartlett PN, Gates DM (1967). “The energy budget of a lizard on a tree trunk.” Ecology, 48, 316-322.

Buckley LB (2008). “Linking traits to energetics and population dynamics to predict lizard ranges in changing environments.” American Naturalist, 171(1), E1 - E19. doi:10.1086/523949 , https://pubmed.ncbi.nlm.nih.gov/18171140/.

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

Riddell EA, Odom JP, Damm JD, Sears MW (2018). “Plasticity reveals hidden resistance to extinction under climate change in the global hotspot of salamander diversity.” Science Advances, 4(4). doi:10.1126/sciadv.aar5471 .

See also

Author

Eric Riddell

Examples

  Qthermal_radiation_absorbed(T_a            = 20,
                              T_g            = 25,
                              epsilon_ground = 0.97,
                              a_longwave     = 0.965)
#> [1] 372.4115