Calculate Reynolds Number

estimate the Reynolds Number, which describes the dynamic properties of the fluid surrounding the animal as the ratio of internal viscous forces (Gates 1980 Biophysical Ecology).

Reynolds_number(u, D, nu)

Arguments

u

is wind speed in m/s

D

is characteristic dimension (e.g., body diameter) (m)

nu

is the kinematic viscosity, ratio of dynamic viscosity to density of the fluid (m^2 s^(-1)), can calculate from DRYAIR or WETAIR

Value

Reynolds number

See also

Examples

Reynolds_number(u=1, D=0.001, nu=1.2)
#> [1] 0.0008333333