Construct a thermal performance curve based on a beta function.

Construct a thermal performance curve based on a beta function (Asbury and Angilletta 2010) .

TPC.beta(T, shift = -1, breadth = 0.1, aran = 0, tolerance = 43, skew = 0.7)

Arguments

T
numeric temperature (C).

shift
numeric mode of the thermal performance curve.

breadth
numeric breadth of the thermal performance curve.

aran
numeric scale performance value. if aran=0, no scaling; if aran=1, include a thermodynamic effect on mean performance.

tolerance
numeric maximal breath (C) of the thermal performance curve.

skew
numeric skewness of the thermal performance curve (0-1).

Value

numeric performance.

References

Asbury DA, Angilletta MJ (2010). “Thermodynamic effects on the evolution of performance curves.” American Naturalist, 176(2), E40--E49. doi: 10.1086/653659 , https://www.journals.uchicago.edu/doi/full/10.1086/653659.

Examples

TPC.beta(T = 0:60, shift = -1, breadth = 0.1, aran = 0, tolerance = 43, skew = 0.7)

#>  [1] 3.803220e-08 2.319533e-06 2.514782e-05 1.343204e-04 4.864521e-04
#>  [6] 1.377095e-03 3.287351e-03 6.923551e-03 1.324542e-02 2.347906e-02
#> [11] 3.911190e-02 6.186767e-02 9.366029e-02 1.365265e-01 1.925380e-01
#> [16] 2.636940e-01 3.517978e-01 4.583191e-01 5.842477e-01 7.299423e-01
#> [21] 8.949814e-01 1.078022e+00 1.276674e+00 1.487399e+00 1.705446e+00
#> [26] 1.924823e+00 2.138334e+00 2.337672e+00 2.513605e+00 2.656239e+00
#> [31] 2.755406e+00 2.801161e+00 2.784431e+00 2.697810e+00 2.536537e+00
#> [36] 2.299662e+00 1.991433e+00 1.622910e+00 1.213837e+00 7.947954e-01
#> [41] 4.096529e-01 1.183447e-01 0.000000e+00 1.000000e-02 1.000000e-02
#> [46] 1.000000e-02 1.000000e-02 1.000000e-02 1.000000e-02 1.000000e-02
#> [51] 1.000000e-02 1.000000e-02 1.000000e-02 1.000000e-02 1.000000e-02
#> [56] 1.000000e-02 1.000000e-02 1.000000e-02 1.000000e-02 1.000000e-02
#> [61] 1.000000e-02