Predicts body temperatures (operative environmental temperature) of an ectotherm in K.

Predicts body temperatures (operative environmental temperature) of an ectotherm in K.

Tb_Gates(A, D, psa_dir, psa_ref, psa_air, psa_g, T_g, T_a, Qabs, epsilon,
  H_L, ef = 1.3, K)

Arguments

A
surface area in m^2

D
characteristic dimension for conduction (m)

psa_dir
proportion surface area exposed to sky (or enclosure)

psa_ref
proportion surface area exposed to ground

psa_air
proportion surface area exposed to air

psa_g
proportion surface area in contact with substrate

T_g
ground surface temperature in K

T_a
ambient air temperature in K

Qabs
Solar and thermal radiation absorbed (W)

epsilon
longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)

H_L
Convective heat transfer coefficient (W m^-2 K^-1)

ef
is the enhancement factor, used to adjust H_L to field condictions (using h_L approximation from Mitchell 1976). Approximated as 1.23 by default, but see Mitchell 1976 for relationship.

K
Thermal conductivity (W K^-1 m^-1 ), K=0.5 W K^-1 m^-1 for naked skin, K=0.15 for insect cuticle ( Galushko et al 2005); conductivity of ground is generally greater than that of animal tissues, so animal thermal conductivity is generally rate limiting step.

Value

operative environmental temperature (K)

Details

Predicts body temperatures (operative environmental temperature) of an ectotherm in K. Uses approximation in Gates (1980, Biophysical ecology). Omits evaporative and metabolic heat loss.

See also

Examples

if (FALSE) {
Tb_Gates(A=1, D=0.001, psa_dir=0.6, psa_ref=0.4, psa_air=0.6, psa_g=0.2, T_g=303, T_a=310, Qabs=800, epsilon=0.95, H_L=10, ef=1.23, K=0.5)
}