Operative Environmental Temperature of an Ectotherm Based on a Variant of Gates (1980)

The function predicts body temperatures (K, operative environmental temperature) of an ectotherm using the approximation in Gates (1980) . The function omits evaporative and metabolic heat loss.

Tb_Gates2(A, D, T_g, T_a, Qabs, V, epsilon)

Arguments

A

numeric surface area (m2).

D

numeric characteristic dimension for conduction (meters).

T_g

numeric ground surface temperature (K).

T_a

numeric ambient air temperature (K).

Qabs

numeric Solar radiation absorbed (W).

V

numeric Wind speed (m s-1).

epsilon

numeric longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980) .

Value

numeric operative environmental temperature (K).

References

Gates DM (1980). Biophysical Ecology. Springer-Verlag, New York, NY, USA.

See also

Examples

  Tb_Gates2(A       = 1, 
            D       = 0.001, 
            T_g     = 300, 
            T_a     = 310, 
            Qabs    = 2, 
            V       = 0.1, 
            epsilon = 1) 
#> [1] 308.0171