Predicts body temperature (operative environmental temperature) of a butterfly in degrees C.

Predicts body temperatures (operative environmental temperatures) of a butterfly in degrees C. Based on Kingsolver (1983, Thermoregulation and flight in Colias butterflies: elevational patterns and mechanistic limitations. Ecology 64: 534–545). Assumes butterfly basks with closed wings perpendicular to solar beam. Source: Buckley LB and Kingsolver JK. 2012. The demographic impacts of shifts in climate means and extremes on alpine butterflies. Functional Ecology https://doi.org/10.1111/j.1365-2435.2012.01969.x and subsequent publications.

Tb_butterfly(T_a, Tg, Tg_sh, u, H_sdir, H_sdif, z, D, delta, alpha,
  r_g = 0.3, shade = FALSE)

Arguments

T_a
is air temperature in C

Tg
is surface temperature in C in the sun

Tg_sh
is surface temperature in C in the shade

u
is wind speed in m/s

H_sdir
is direct solar radiation flux in W/m^2

H_sdif
is diffuse solar radiation flux in W/m^2

z
is solar zenith angle in degrees

D
is thoracic diameter in cm

delta
is thoracic fur thickness in mm

alpha
is wing solar absorbtivity as a proportion. Range for Colias is 0.4 to 0.7.

r_g
is substrate solar reflectivity (proportion), see Kingsolver (1983)

shade
is whether body temperature should be calculate in sun (FALSE) or shade (TRUE)

Value

predicted body (operative environmental) temperature (C)

Details

Predicts body temperature (operative environmental temperature) of a butterfly in degrees C.

Examples

if (FALSE) {
Tb_butterfly(T_a=25, Tg=25, Tg_sh=20, u=0.4, H_sdir=300, H_sdif=100, z=30, D=0.36, delta=1.46, alpha=0.6, r_g=0.3)
}