Operative Environmental Temperature of a Limpet

Predicts body temperatures (operative environmental temperatures) of a limpet in C.

Based on Denny and Harley (2006) .

Tb_limpet(T_a, T_r, L, H, I, u, psi, c, position = "anterior")

Arguments

T_a

numeric air temperature in C.

T_r

numeric rock surface temperature in C in the sunlight.

L

numeric limpet length (anterior/posterior axis) (m).

H

numeric limpet height (dorsal/ventral axis) (m).

I

numeric solar irradiance W / m2.

u

numeric wind speed (m / s).

psi

numeric solar zenith angle (degrees): can be calculated from zenith_angle function.

c

numeric fraction of the sky covered by cloud.

position

character direction of the limpet that is facing upwind. Options are "anterior", "posterior", and "broadside".

Value

numeric predicted body (operative environmental) temperature (C).

Details

The original equation uses a finite-difference approach where they divide the rock into series of chunks, and calculate the temperature at each node to derive the conductive heat. For simplification, here it takes the rock temperature as a parameter, and conductive heat is calculated by the product of the area, thermal conductivity of rock and the difference in temperatures of the rock and the body.

Limpets are simulated as cones following and using solar emissiviity values from Campbell and Norman (1998) .

The area of the limpet's shell (m2) is projected in the direction at which sunlight strikes the organism Pennell and Deignan (1989) .

Air conductivity values (W / (m-1 K-1)) are calculated following Denny and Harley (2006) .

References

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

Denny MW, Harley CDG (2006). “Hot limpets: predicting body temperature in a conductance-mediated thermal system.” Journal of Experimental Biology, 209(13), 2409–2419. ISSN 0022-0949, doi: 10.1242/jeb.02356 , https://journals.biologists.com/jeb/article-pdf/209/13/i/1255483/isi-iii.pdf.

Pennell S, Deignan J (1989). “Computing the Projected Area of a Cone.” SIAM Review, 31, 299-302. https://epubs.siam.org/doi/10.1137/1031052.

See also

Examples

  Tb_limpet(
    T_a = 25, 
    T_r = 30, 
    L = 0.0176, 
    H = 0.0122, 
    I = 1300, 
    u = 1, 
    psi = 30, 
    c = 1, 
    position = "anterior")