Operative Environmental Temperature of a Lizard

The function estimates body temperature (C, operative environmental temperature) of a lizard based on Campbell and Norman (1998) . The function was designed for Sceloporus lizards and described in Buckley (2008) .

Tb_lizard(
  T_a,
  T_g,
  u,
  svl,
  m,
  psi,
  rho_S,
  elev,
  doy,
  sun = TRUE,
  surface = TRUE,
  a_S = 0.9,
  a_L = 0.965,
  epsilon_s = 0.965,
  F_d = 0.8,
  F_r = 0.5,
  F_a = 0.5,
  F_g = 0.5
)

Arguments

T_a

numeric air temperature (C).

T_g

numeric surface temperature (C).

u

numeric wind speed (m s-1).

svl

numeric lizard snout vent length (mm).

m

numeric lizard mass (g); note that it can be estimated as in mass_from_length: 3.55 x 10-5 x length3

psi

numeric solar zenith angle (degrees).

rho_S

numeric surface albedo (proportion). ~ 0.25 for grass, ~ 0.1 for dark soil, > 0.75 for fresh snow (Campbell and Norman 1998) .

elev

numeric elevation (m).

doy

numeric day of year (1-366).

sun

logical indicates whether lizard is in sun (TRUE) or shade (FALSE).

surface

logical indicates whether lizard is on ground surface (TRUE) or above the surface (FALSE, e.g. in a tree).

a_S

numeric lizard solar absorptivity (proportion), a_S = 0.9 (Gates 1980) (Table 11.4).

a_L

numeric lizard thermal absorptivity (proportion), a_L = 0.965 (Bartlett and Gates 1967) .

epsilon_s

numeric surface emissivity of lizards (proportion), epsilon_s = 0.965 (Bartlett and Gates 1967) .

F_d

numeric the view factor between the surface of the lizard and diffuse solar radiation (proportion). i.e., the portion of the lizard surface that is exposed to diffuse solar radiation (Bartlett and Gates 1967) .

F_r

numeric the view factor between the surface of the lizard and reflected solar radiation (proportion).

F_a

numeric the view factor between the surface of the lizard and atmospheric radiation (proportion).

F_g

numeric the view factor between the surface of the lizard and ground thermal radiation (proportion).

Value

T_e numeric predicted body (operative environmental) temperature (C).

Details

The proportion of radiation that is direct is determined following Sears et al. (2011) .

Boundary conductance uses a factor of 1.4 to account for increased convection (Mitchell 1976) .

References

Bartlett PN, Gates DM (1967). “The energy budget of a lizard on a tree trunk.” Ecology, 48, 316-322.

Buckley LB (2008). “Linking traits to energetics and population dynamics to predict lizard ranges in changing environments.” American Naturalist, 171(1), E1 - E19. doi:10.1086/523949 , https://pubmed.ncbi.nlm.nih.gov/18171140/.

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

Gates DM (1980). Biophysical Ecology. Springer-Verlag, New York, NY, USA.

Mitchell JW (1976). “Heat transfer from spheres and other animal forms.” Biophysical Journal, 16(6), 561-569. ISSN 0006-3495, doi:10.1016/S0006-3495(76)85711-6 , https://www.sciencedirect.com/science/article/pii/S0006349576857116.

Sears MW, Raskin E, Angilletta Jr. MJ (2011). “The World Is not Flat: Defining Relevant Thermal Landscapes in the Context of Climate Change.” Integrative and Comparative Biology, 51(5), 666-675. ISSN 1540-7063, doi:10.1093/icb/icr111 , https://academic.oup.com/icb/article-pdf/51/5/666/1757893/icr111.pdf.

See also

Examples

  Tb_lizard(T_a       = 25, 
            T_g       = 30, 
            u         = 0.1, 
            svl       = 60, 
            m         = 10, 
            psi       = 34, 
            rho_S     = 0.24, 
            elev      = 500, 
            doy       = 200, 
            sun       = TRUE, 
            surface   = TRUE, 
            a_S   = 0.9, 
            a_L   = 0.965, 
            epsilon_s = 0.965, 
            F_d       = 0.8, 
            F_r       = 0.5, 
            F_a       = 0.5, 
            F_g       = 0.5)
#> Error in Tb_lizard(T_a = 25, T_g = 30, u = 0.1, svl = 60, m = 10, psi = 34,     rho_S = 0.24, elev = 500, doy = 200, sun = TRUE, surface = TRUE,     a_S = 0.9, a_L = 0.965, epsilon_s = 0.965, F_d = 0.8, F_r = 0.5,     F_a = 0.5, F_g = 0.5): unused arguments (rho_S = 0.24, a_S = 0.9, a_L = 0.965)