Predicts body temperature (operative environmental temperature) of a mussel in °C.

Predicts body temperature of a mussel in °C. Implements a steady‐state model, which assumes unchanging environmental conditions. Based on Helmuth 1998, INTERTIDAL MUSSEL MICROCLIMATES: PREDICTING THE BODY TEMPERATURE OF A SESSILE INVERTEBRATE

Tb_mussel(L, H, T_a, T_g, S, k_d, u, psi, evap = FALSE, cl, group = "solitary")

Arguments

L
mussel length (anterior/posterior axis) (m)

H
mussel height (dorsal/ventral axis) (m), reasonable to assume 0.5*L

T_a
air temperature (°C)

T_g
ground temperature (°C)

S
direct solar flux density (W/m2)

k_d
diffuse fraction, proportion of solar radiation that is diffuse

u
wind speed (m/s)

psi
solar zenith angle (degrees): can be calculated from zenith_angle function

evap
Are mussels gaping to evaporatively cool? TRUE of FALSE (default), If TRUE, assumes constant mass loss rate of 5 percent of initial body mass per hour

cl
fraction of the sky covered by cloud

group
options are "aggregated": mussels living in beds, "solitary": mussels individuals, anterior or posterior end facing upwind, and "solitary_valve": solitary individuals, valve facing upwind

Value

predicted body temperature (°C)

Details

Predicts body temperature of a mussel in °C.

See also

Examples

if (FALSE) {
Tb_mussel(
  L = 0.1, 
  H = 0.05, 
  T_a = 25, 
  T_g = 30, 
  S=500, 
  k_d=0.2, 
  u = 2, 
  psi = 30, 
  evap = FALSE, 
  cl = 0.5, 
  group = "solitary")
}