Humid Operative Temperature

Calculate humid operative temperature, using adaptation of Campbell and Norman (1998) (Riddell et al. 2018) .

Tb_salamander_humid(r_i, r_b, D, T_a, elev, e_a, e_s, Qabs, epsilon = 0.96)

Arguments

r_i
numeric internal (skin) resistance (s cm-1).

r_b
numeric boundary layer resistance (s cm-1).

D
numeric body diameter (meters), (diameter = 0.0016*log(mass) + 0.0061 for mass(g)).

T_a
numeric ambient temperature (C).

elev
numeric elevation (m).

e_a
numeric actual vapor pressure (kPa).

e_s
numeric saturation vapor pressure (kPa).

Qabs
numeric Solar and thermal radiation absorbed (W).

epsilon
numeric emissivity of salamander skin.

Value

numeric humid operative temperature (C).

References

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

Riddell EA, Odom JP, Damm JD, Sears MW (2018). “Plasticity reveals hidden resistance to extinction under climate change in the global hotspot of salamander diversity.” Science Advances, 4(4). doi: 10.1126/sciadv.aar5471 , https://doi.org/10.1126/sciadv.aar5471.

See also

Author

Eric Riddell

Examples

  Tb_salamander_humid(r_i     = 4,
                      r_b     = 1,
                      D       = 0.01,
                      T_a     = 20,
                      elev    = 500,
                      e_a     = 2.0,
                      e_s     = 2.5,
                      Qabs    = 400,
                      epsilon = 0.96)

#> [1] 19.25887