Predict Body Temperature (Operative Environmental Temperature) of a Marine Snail in C

Predicts body temperature of a snail in C. Implements a steady-state model, which assumes unchanging environmental conditions. Based on (Iacarella and Helmuth 2012) . Body temperature and desiccation constrain the activity of Littoraria irrorata within the Spartina alterniflora canopy. Function was provided by the Helmuth lab and is a simplified version of model in publication.

Tb_snail(temp, l, solar, WS, CC, WL, WSH)

Arguments

temp
numeric air temperature (C).

l
numeric snail length (m).

solar
numeric direct solar flux density (W m-2).

WS
numeric wind speed (m s-1).

CC
numeric fraction of the sky covered by cloud (0-1).

WL
numeric water loss rate (kg s-1), 5 percent loss of body mass over one hour is a reasonable maximum level (Helmuth 1999) .

WSH
numeric wind sensor height (m).

Value

numeric predicted body temperature (C).

Details

Thermal radiative flux is calculated following Helmuth (1998) , Helmuth (1999) , and Idso and Jackson (1969) .

References

Helmuth B (1999). “Thermal biology of rocky intertidal mussels: quantifying body temperatures using climatological data.” Ecology, 80(1), 15--34. doi: 10.1890/0012-9658(1999)080[0015:TBORIM]2.0.CO;2 , https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/0012-9658%281999%29080%5B0015%3ATBORIM%5D2.0.CO%3B2.

Helmuth BST (1998). “Intertidal Mussel Microclimates: Predicting the Body Temperature of a Sessile Invertebrate.” Ecological Monographs, 68(1), 51--74. ISSN 00129615, http://www.jstor.org/stable/2657143.

Iacarella J, Helmuth B (2012). “Body temperature and desiccation constrain the activity of Littoraria irrorata within the Spartina alterniflora canopy.” Journal of Thermal Biology, 37(1). doi: 10.1016/j.jtherbio.2011.10.003 , https://www.researchgate.net/publication/238501940_Body_temperature_and_desiccation_constrain_the_activity_of_Littoraria_irrorata_within_the_Spartina_alterniflora_canopy.

Idso SB, Jackson RD (1969). “Thermal radiation from the atmosphere.” Journal of Geophysical Research (1896-1977), 74(23), 5397--5403. doi: 10.1029/JC074i023p05397 , https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/JC074i023p05397.

See also

Author

Brian Helmuth Lab

Examples

  Tb_snail(temp  = 25, 
           l     = 0.012, 
           solar = 800, 
           WS    = 1, 
           CC    = 0.5, 
           WL    = 0, 
           WSH   = 10)

#> [1] 38.79568