TrenchR: Tools for Microclimate and Biophysical Ecology

Tools for translating environmental change into organismal response. Microclimate models to vertically scale weather station data to organismal heights. The biophysical modeling tools include both general models for heat flows and specific models to predict body temperatures for a variety of ectothermic taxa. Additional functions model and temporally partition air and soil temperatures and solar radiation. Utility functions estimate the organismal and environmental parameters needed for biophysical ecology. 'TrenchR' focuses on relatively simple and modular functions so users can create transparent and flexible biophysical models. Many functions are derived from Gates (1980) doi:10.1007/978-1-4612-6024-0 and Campbell and Norman (1988) <isbn:9780387949376>.

See also

Author

Maintainer: Lauren Buckley lbuckley@uw.edu (ORCID)

Authors:

Bryan Briones Ortiz

Isaac Caruso

Aji John (ORCID)

Ofir Levy (ORCID)

Abigail Meyer

Eric Riddell (ORCID)

Yutaro Sakairi

Juniper Simonis (ORCID)

Other contributors:

Brian Helmuth (ORCID) [contributor]