Calculate Actual Vapor Pressure from Dewpoint Temperature

Calculate actual vapor pressure from dewpoint temperature based on Stull (2000); Eric Riddell and Sears (2018) .

actual_vapor_pressure(T_dewpoint)

Arguments

T_dewpoint

numeric dewpoint temperature (C).

Value

numeric actual vapor pressure (kPa).

References

Eric Riddell JPOJDD, Sears MW (2018). “Plasticity reveals hidden resistance to extinction under climate change in the global hotspot of salamander diversity.” Science Advances, 4(4). doi: 10.1126/sciadv.aar5471 , https://www.researchgate.net/publication/326332138_Plasticity_reveals_hidden_resistance_to_extinction_under_climate_change_in_the_global_hotspot_of_salamander_diversity.

Stull RB (2000). Meteorology for Scientists and Engineers. Brooks Cole. ISBN 978-0534372149.

See also

Author

Eric Riddell

Examples

  actual_vapor_pressure(T_dewpoint = 20)
#> [1] 2.393477