Estimate temperature at a specified height

calculate temperature (C) at a specified height (m). Estimates a three segment velocity and temperature profile based on user-specified, experimentally determined values for 3 roughness heights and reference heights. Multiple heights are appropriate in heterogenous areas with, for example, a meadow, bushes, and rocks. Implements the MICROSEGMT routine from NicheMapR as described in Kearney and Porter 2016. Source: Kearney MR and Porter WP. 2016. NicheMapR – an R package for biophysical modelling: the microclimate model. Ecography 40:664-674.

air_temp_profile_segment(T_r, u_r, zr, z0, z, T_s)

Arguments

T_r

is a vector of temperature at the 3 reference heights in C.

u_r

is a vector of wind speeds at the 3 reference heights in m/s.

zr

is a vector of 3 reference heights in m.

z0

is a vector of 3 experimentally determined roughness heights in m.

z

is height to scale to in m.

T_s

is surface temperatures in C.

See also

Examples

air_temp_profile_segment(
 T_r=c(25,22,20),
 u_r=c(0.01,0.025,0.05), 
 zr=c(0.05,0.25,0.5), 
 z0=c(0.01,0.15,0.2), 
 z=0.3, 
 T_s=27)