Estimate the boundary layer resistance

Estimate the boundary layer resistance

boundary_layer_resistance(T_a, e_s, e_a, elev, D, u = NA)

Arguments

T_a
air temperature (K)

e_s
saturation vapor pressure (kPa)

e_a
actual vapor pressure (kPa)

elev
elevation (m)

D
characteristic dimension (e.g., body diameter) (m), ##diameter = 0.0016*log(mass) + 0.0061 for mass(g) #empirical formula for salamander diameter, Riddell et al. 2017

u
is wind speed in m/s, if not provided assume free convection; if provided, use forced convection if appropriate

Value

boundary layer resistance (s cm^-1)

Details

This function allows you to estimate boundary layer resistance under free convection. Based on the function in Riddell et al. 2017. Source: Riddell EA. 2017. Physical calculations of resistance to water loss improve predictions of species range models. Ecological Monographs 87: 21-23.

See also

Examples

if (FALSE) {
boundary_layer_resistance(T_a=293, e_s=2.4, e_a=2.5, elev=500, D=0.007, u=2)
}