Calculate degree days

calculate degree days using single or double sine wave and single or double triangulation approximation. Source: http://ipm.ucanr.edu/WEATHER/ddfigindex.html. Double methods assume symmetry, that is that next day's thermal minima is equal to previous day. Double Sine wave approximation of degree days from Allen JC. 1976. A Modified Sine Wave Method for Calculating Degree Days. Environmental Entomology 5:388–396.

degree_days(T_min, T_max, LDT = NA, UDT = NA, method = "single.sine")

Arguments

T_min

Minimum temperature of the day (C)

T_max

Maximum temperature of the day (C)

LDT

lower developmental threshold (C).

UDT

upper developmental threshold (C).

method

type of method being used. Current choices: "single.sine","double.sine", "single.triangulation" or "double.triangulation".

Value

degree days (C)

See also

Examples

degree_days(T_min=7, T_max=14, LDT=12, UDT=33, method="single.sine")
#> [1] 0.47
degree_days(T_min=7, T_max=14, LDT=12, UDT=33, method="single.triangulation")
#> [1] 0.29