Direct Solar Radiation

The function estimates direct solar radiation (W/m2) based on latitude, day of year, elevation, and time. The function uses two methods (McCullough and Porter 1971; Campbell and Norman 1998) compiled in Tracy et al. (1983) .

direct_solar_radiation(lat, doy, elev, t, t0, method = "Campbell 1977")

Arguments

lat

numeric latitude (degrees).

doy

numeric day of year (1-366).

elev

numeric elevation (m).

t

numeric local time (decimal hours).

t0

numeric time of local noon (decimal hours), can be estimated using solar_noon.

method

character method for estimating direct solar radiation, options: "Campbell 1977" (default), "Gates 1962".

Value

numeric direct solar radiation (W/m2).

References

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

McCullough EC, Porter WP (1971). “Computing Clear Day Solar Radiation Spectra for the Terrestrial Ecological Environment.” Ecology, 52(6), 1008-1015. doi:10.2307/1933806 .

Tracy CR, Hammond KA, Lechleitner RA, II WJS, Thompson DB, Whicker AD, Williamson SC (1983). “Estimating clear-day solar radiation: an evaluation of three models.” Journal of Thermal Biology, 8(3), 247-251. doi:10.1016/0306-4565(83)90003-7 .

See also

Examples

  direct_solar_radiation(lat    = 47.61, 
                         doy    = 112, 
                         elev   = 1500, 
                         t      = 9, 
                         t0     = 12, 
                         method = "Campbell 1977")
#> [1] 656.1941