Estimates temperature across hours using sine and square root functions

estimate temperature for a specified hour using sine and square root functions. Source: Cesaraccio C et al. 2001. An improved model for determining degree-day values from daily temperature data. International Journal of Biometeorology 45:161-169.

diurnal_temp_variation_sinesqrt(t, tr, ts, T_max, T_min, T_minp)

Arguments

t

hour or hours for temperature estimate

tr

sunrise hour (0-23)

ts

sunset hour (0-23)

T_max

maximum temperature of current day (C)

T_min

minimum temperature of current day (C)

T_minp

minimum temperature of following day (C)

Value

temperature (C) at a specified hour

Details

Estimates temperature across hours using sine and square root functions

See also

Examples

diurnal_temp_variation_sinesqrt(t=8, tr=6, ts=18, T_max=30, T_min=10, T_minp=12)
#> [1] 17.65367