Estimate radiation (three parts - Direct, Diffused and Reflected)

Estimate direct, diffuse, and reflected components of solar radiation in W m^-2 using the model in Campbell & Norman (1998).

estimate_radiation(doy, psi, tau, elev, rho = 0.7)

Arguments

doy
is the day of year

psi
is the zenith angle in radians

tau
atmospheric transmissivity (proportion), which is ratio of global solar radiation at ground level to extra-terrestrial solar radiation

elev
is elevation (m)

rho
is the albedo as a proportion

Value

radiation components - direct, diffused and reflected (W/m^2)

See also

Examples

if (FALSE) {
estimate_radiation(doy=112, psi=1, tau=0.6, elev=1500, rho=0.7)
}