Estimate the Heat Transfer Coefficient Using a Spherical Approximation

Estimate the heat transfer coefficient for various taxa. Approximates forced convective heat transfer for animal shapes using the convective relationship for a sphere (Mitchell 1976) .

heat_transfer_coefficient_approximation(V, D, K, nu, taxon = "sphere")

Arguments

V

numeric air velocity (m s-1).

D

numeric characteristic dimension (e.g., diameter or snout-vent length) (meters).

K

numeric thermal conductivity of air (W m-2 K-1), can calculate using DRYAIR or WETAIR.

nu

numeric kinematic Viscosity of air (m2 s-1), can calculate using DRYAIR or WETAIR.

taxon

character of which class of organism, current choices: "sphere", "frog", "lizard", "flyinginsect", "spider".

Value

numeric heat transfer coefficient, H_L (W m-2 K-1).

References

Mitchell JW (1976). “Heat transfer from spheres and other animal forms.” Biophysical Journal, 16(6), 561--569. ISSN 0006-3495, doi: 10.1016/S0006-3495(76)85711-6 , https://www.sciencedirect.com/science/article/pii/S0006349576857116.

See also

Examples

 heat_transfer_coefficient_approximation(V     = 3,
                                         D     = 0.05,
                                         K     = 25.7 * 10^(-3),
                                         nu    = 15.3 * 10^(-6),
                                         taxon = "sphere")
#> [1] 43.37924