Diffuse Fraction for Partitioning Solar Radiation

The function partitions solar radiation (W m-2) into direct and diffuse components by estimating the diffuse fraction (k_d). The function uses the models presented in Wong and Chow (2001) .

partition_solar_radiation(method, kt, lat = NA, sol.elev = NA)

Arguments

method

character method to use for estimating the diffuse fraction, currently available: "Liu_Jordan", "Orgill_Hollands", "Erbs", "Olyphant", "Spencer", "Reindl-1", "Reindl-2", "Lam_Li".

kt

numeric the clearness index (dimensionless), which is the ratio of the global solar radiation measured at the surface to the total solar radiation at the top of the atmosphere. (0-1)

lat

numeric latitude (degrees). Needed only if method is "Spencer".

sol.elev

numeric the solar elevation angles (degrees). Needed only if method is "Reindl-2".

Value

numeric diffuse fraction.

References

Wong LT, Chow WK (2001). “Solar radiation model.” Applied Energy, 69(3), 191-224. ISSN 0306-2619, doi:10.1016/S0306-2619(01)00012-5 , https://www.sciencedirect.com/science/article/pii/S0306261901000125.

See also

Examples

  partition_solar_radiation(method   = "Erbs", 
                            kt       = 0.5, 
                            lat      = 40, 
                            sol.elev = 60)
#> [1] 0.65915