Organism Silhouette Area

Estimate the projected (silhouette) area as a portion of the surface area of the organism. Estimates the projected area as a function of zenith angle.

proportion_silhouette_area(z, taxon, raz = 0, posture = "prostrate")

Arguments

z
numeric zenith angle in degrees between 0 and 360.

taxon
character Organism taxon. Current choices are "lizard", "frog", and "grasshopper".

raz
numeric relative solar azimuth angle (in degrees). Required if taxon = "lizard". This is the horizontal angle of the sun relative to the head and frontal plane of the lizard and options currently include 0 (in front), 90 (to side), and 180 (behind) degrees.

posture
character value describing posture. Required if taxon = "lizard". Options include "prostrate" (default) and "elevated".

Value

numeric silhouette area as a proportion.

Details

Relationships come from

Lizards: Muth (1977) .

Frogs: Tracy (1976) .

Grasshoppers: Anderson et al. (1979) .

References

Anderson RV, Tracy CR, Abramsky Z (1979). “Habitat Selection in Two Species of Short-Horned Grasshoppers. The Role of Thermal and Hydric Stresses.” Oecologia, 38(3), 359--374. ISSN 00298549, 14321939, http://www.jstor.org/stable/4215792.

Muth A (1977). “Thermoregulatory Postures and Orientation to the Sun: A Mechanistic Evaluation for the Zebra-Tailed Lizard, Callisaurus draconoides.” Copeia, 4, 710--720.

Tracy CR (1976). “A Model of the Dynamic Exchanges of Water and Energy between a Terrestrial Amphibian and Its Environment.” Ecological Monographs, 46(3), 293--326. doi: 10.2307/1942256 , https://esajournals.onlinelibrary.wiley.com/doi/abs/10.2307/1942256.

See also

Examples

  proportion_silhouette_area(z     = 60,   
                             taxon = "frog")

#> [1] 0.2902208

  proportion_silhouette_area(z     = 60, 
                             taxon = "grasshopper")

#> [1] 0.0862

  proportion_silhouette_area(z       = 60, 
                             taxon   = "lizard", 
                             posture = "prostrate", 
                             raz     = 90)

#> [1] 0.329857

  proportion_silhouette_area(z       = 60, 
                             taxon   = "lizard", 
                             posture = "elevated", 
                             raz     = 180)

#> [1] 0.176596