Organism Silhouette Area using Shape Approximations

Estimate the projected (silhouette) area as a portion of the surface area of the organism. Estimate the projected area as a function of the dimensions and the angle between the solar beam and the longitudinal axis of the solid, using Figure 11.6 in Campbell and Norman (1998) .

proportion_silhouette_area_shapes(shape, theta, h, d)

Arguments

shape
character Which shape to approximate an organism. Shapes are assumed to be prolate or have the longest axis parallel with the ground. Current choices are "spheroid", "cylinder flat ends", and "cylinder hemisphere ends".

theta
numeric angle between the solar beam and the longitudinal axis in degrees.

h
numeric height (long axis in m). Cross section length for spheroid.

d
numeric diameter (short axis in m). Cross section length for spheroid.

Value

numeric silhouette area as a proportion.

References

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

See also

Examples

  proportion_silhouette_area_shapes(shape = "spheroid", 
                                    theta = 60,  
                                    h     = 0.01,  
                                    d     = 0.001)

#> [1] 0.2748564

  proportion_silhouette_area_shapes(shape = "cylinder flat ends",  
                                    theta = 60,  
                                    h     = 0.01, 
                                    d     = 0.001)

#> [1] 0.2744423

  proportion_silhouette_area_shapes(shape = "cylinder hemisphere ends",  
                                    theta = 60,  
                                    h     = 0.01, 
                                    d     = 0.001)

#> [1] 0.2733313