Estimate soil thermal conductivity

estimate soil thermal conductivity in W m^-1 K^-1 using the methods of de Vries (1963, The Physics of Plant Environments, Ch2 in Environmental Control of Plant Growth).

soil_conductivity(x, lambda, g_a)

Arguments

x

is a vector of volume fractions of soil constituents (e.g., clay, quartz, minerals other than quartz, organic matter, water, air). The volume fractions should sum to 1. Note that x and lambda values in the example correspond to these soil constituents.

lambda

is a vector of the thermal conductivities (W m^-1 K^-1) of the soil constituents.

g_a

is a shape factor on soil particles. The soil particles are assumed to be ellipsoids with axes g_a, g_b, and g_c, where g_a +g_b +g_c=1 and g_a=g_b. de Vries 1952 suggests g_a=g_b=0.125.

Value

soil thermal conductivity (W m^-1 K^-1)

See also

Author

Joseph Grigg

Examples

soil_conductivity(
  x=c(0.10,0.40,0.11,0.01,0.2, 0.18), 
  lambda=c(0.10,0.40,0.11,0.01,0.2, 0.18), 
  g_a=0.125)