Calculate Soil Temperature in C using ODEs

Function called to calculate soil temperature in C from W.A. Beckman and Porter (1973) . Calls soil_temperature_function, which uses ODE to calculate soil profile. This is the primary function to call to estimate soil temperature. Uses equations from (deVries 1963)

soil_temperature(
  z_r.intervals = 12,
  z_r,
  z,
  T_a,
  u,
  Tsoil0,
  z0,
  SSA,
  TimeIn,
  H,
  water_content = 0.2,
  air_pressure,
  rho_so = 1620,
  shade = FALSE
)

Arguments

z_r.intervals
numeric the number of intervals in the soil profile to calculate, defaults to 12.

z_r
numeric reference height in0 (m).

z
numeric interval of the soil profile to return (1 to z_r.intervals).

T_a
numeric vector of air temperature (degrees C), Note: missing values will be linearly interpolated.

u
numeric vector of wind speed (m/s).

Tsoil0
numeric initial soil temperature (degrees C).

z0
numeric surface roughness (m).

SSA
numeric solar absorptivity of soil surface as a fraction.

TimeIn
numeric vector of time periods for the model.

H
numeric vector of solar radiation (W/m2).

water_content
numeric percent water content (percent).

air_pressure
numeric air pressure (kPa).

rho_so
numeric particle density of soil.

shade
logical whether or not soil temperature should be calculated in the shade.

Value

numeric soil temperature (C).

References

deVries DA (1963). “Thermal Properties of Soils.” W. R. Van Wijk's Physics of Plant Environment, Chapter 2. ISSN 19640701041, doi: 10.1002/qj.49709038628 , https://doi.org/10.1002/qj.49709038628.

W.A. Beckman JWM, Porter WP (1973). “Thermal Model for Prediction of a Desert Iguana㤼㸲s Daily and Seasonal Behavior.” J. Heat Transfer, 95(2), 257--262. doi: 10.1115/1.3450037 , https://asmedigitalcollection.asme.org/heattransfer/article-abstract/95/2/257/416609/Thermal-Model-for-Prediction-of-a-Desert-Iguana-s?redirectedFrom=fulltext.

See also

Author

Joseph Grigg

Examples

  set.seed(123)
  temp_vector       <- runif(48, min = -10, max = 10)
  wind_speed_vector <- runif(48, min = 0, max = 0.4)
  time_vector       <- rep(1:24, 2)
  solrad_vector     <- rep(c(rep(0, 6), 
                             seq(10, 700, length.out = 6), 
                             seq(700, 10, length.out = 6), 
                             rep(0, 6)),
                           2)

  soil_temperature(z_r.intervals = 12, 
                   z_r           = 1.5, 
                   z             = 2, 
                   T_a           = temp_vector, 
                   u             = wind_speed_vector, 
                   Tsoil0        = 20, 
                   z0            = 0.02, 
                   SSA           = 0.7, 
                   TimeIn        = time_vector, 
                   H             = solrad_vector, 
                   water_content = 0.2, 
                   air_pressure  = 85, 
                   rho_so        = 1620, 
                   shade         = FALSE)

#>  [1] 20.000000 12.339950 11.180040  9.286421  9.519004  9.276450  5.876114
#>  [8]  5.850459 10.020513 12.991114 15.664699 22.013450 26.230581 28.463296
#> [15] 28.649168 23.767555 23.091116 15.606473 10.403332  8.170456  9.723145
#> [22]  9.114185  7.811254  6.684010  7.773036  6.070261  5.836625  4.626048
#> [29]  4.301875  2.632974  1.038409  3.946260  7.266697 10.732009 15.272042
#> [36] 17.527402 22.648567 28.053649 24.650392 20.349976 16.396396 11.772210
#> [43]  9.453015  7.779332  5.773844  4.137864  2.554147  2.289415