Core Function Called to Solve Equation for Soil Temperature

Function called by soil_temperature_function to solve equation for soil temperature from W.A. Beckman and Porter (1973) .

soil_temperature_equation(L, rho_a, c_a, k, V_inst, z_r, z0, T_inst, T_s)

Arguments

L
numeric Monin-Obukhov length, a measure of the instability of heat flow (see W.A. Beckman and Porter (1973) ).

rho_a
numeric density of air (kg m-3).

c_a
numeric specific heat of air (J kg-1 K-1).

k
numeric von Karman's constant.

V_inst
numeric instantaneous wind speed (m s-1).

z_r
numeric reference height (m).

z0
numeric surface roughness (m).

T_inst
numeric instantaneous air temperature (K).

T_s
numeric initial soil surface temperature (C).

Value

soil temperature function

References

W.A. Beckman JWM, Porter WP (1973). “Thermal Model for Prediction of a Desert Iguana㤼㸲s Daily and Seasonal Behavior.” J. Heat Transfer, 95(2), 257--262. doi: 10.1115/1.3450037 , https://asmedigitalcollection.asme.org/heattransfer/article-abstract/95/2/257/416609/Thermal-Model-for-Prediction-of-a-Desert-Iguana-s?redirectedFrom=fulltext.

See also

Author

Joseph Grigg

Examples

soil_temperature_equation(L      = -10, 
                          rho_a  = 1.177, 
                          c_a    = 1006, 
                          k      = 0.41, 
                          V_inst = 0.3, 
                          z_r    = 1.5, 
                          z0     = 0.02, 
                          T_inst = 265, 
                          T_s    = 20)

#> [1] 1340.707