Solve equation for soil temperature

Function called by soil_temp_noint to solve equation for soil temperature from Beckman et al. (1973, Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior). The function represents the integrand in the equation. It is not intended to be called directly.

soil_temperature_integrand(x, L, z0)

Arguments

x
is a vector of volume fractions of soil constituents (e.g., clay, quartz, minerals other than quartz, organic matter, water, air). The volume fractions should sum to 1. Note that x and lambda values in the example correspond to these soil constituents.

L
is the Monin-Obukhov length, a measure of the instability of heat flow (see Beckman et al. 1973)

z0
is surface rougness in m

Value

integrand for soil temperature function

Details

Function called by soil_temp_equation to solve equation for soil temperature.

See also

Examples

if (FALSE) {
soil_temperature_integrand(x=c(0.10,0.40,0.11,0.01,0.2, 0.18), L=-10,z0=0.2)
}