Organism Surface Area from Length

This function estimates surface area (m2) from length (m) by approximating the animal's body as a rotational ellipsoid with half the body length as the semi-major axis.

surface_area_from_length(l)

Arguments

l

numeric length (m).

Value

numeric surface area (m2).

Details

Following Samietz et al. (2005) and Lactin and Johnson (1998) .

References

Lactin DJ, Johnson DL (1998). “Convective heat loss and change in body temperature of grasshopper and locust nymphs: Relative importance of wind speed, insect size and insect orientation.” Journal of Thermal Biology, 23(1), 5-13. ISSN 0306-4565, doi:10.1016/S0306-4565(97)00037-5 , https://www.sciencedirect.com/science/article/pii/S0306456597000375.

Samietz J, Salser MA, Dingle H (2005). “Altitudinal variation in behavioural thermoregulation: local adaptation vs. plasticity in California grasshoppers.” Journal of Evolutionary Biology, 18(4), 1087-1096. doi:10.1111/j.1420-9101.2005.00893.x .

See also

Examples

  surface_area_from_length(l = 0.04)
#> [1] 0.00187877