Allometry: Surface Area from Mass

Estimate surface area (m2) from mass (g) for one of a variety of taxa.

surface_area_from_mass(M, taxon)

Arguments

M

numeric vector of mass (g).

taxon

character taxonomic classification of organism, current choices: "lizard", "salamander", "frog", "insect".

Value

numeric surface area (m2).

Details

All models follow (SA = a Mb) with mass in grams and surface area in meters2.

Lizards (Norris 1965; Porter and James 1979; Roughgarden 1981; OConnor 1999; Fei et al. 2012) :

\(a = 0.000314 \pi\)
\(b = 2/3\)

Salamanders (Whitford and Hutchison 1967; Riddell et al. 2017) :

\(a = 0.000842\)
\(b = 0.694\)

Frogs (McClanahan and Baldwin 1969) :

\(a = 0.00099\)
\(b = 0.56\)

Insects (Lactin et al. 1997) .

\(a = 0.0013\)
\(b = 0.8\)

References

Fei T, Skidmore AK, Venus V, Wang T, Schlerf M, Toxopeus B, van Overjijk S, Bian M, Liu Y (2012). “A body temperature model for lizards as estimated from the thermal environment.” Journal of Thermal Biology, 37(1), 56-64. ISSN 0306-4565, doi: 10.1016/j.jtherbio.2011.10.013 , https://www.sciencedirect.com/science/article/pii/S0306456511001513.

Lactin DJ, L. D, Johnson (1997). “Response of body temperature to solar radiation in restrained nymphal migratory grasshoppers (Orthoptera: Acrididae): influences of orientation and body size.” Physiological Entomology, 22(2), 131-139. doi: 10.1111/j.1365-3032.1997.tb01150.x , https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1365-3032.1997.tb01150.x, https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-3032.1997.tb01150.x.

McClanahan L, Baldwin R (1969). “Rate of water uptake through the integument of the desert toad, Bufo punctatus.” Comparative Biochemistry and Physiology, 28(1), 381-389. ISSN 0010-406X, doi: 10.1016/0010-406X(69)91351-6 , https://www.sciencedirect.com/science/article/pii/0010406X69913516.

Norris KS (1965). “Color adaptation in desert reptiles and its thermal relationships.” In Symposium on Lizard Ecology, 162-– 229. U. Missouri Press.

OConnor M (1999). “Physiological and ecological implications of a simple model of heating and cooling in reptiles.” Journal of Thermal Biology, 24, 113-136.

Porter WP, James FC (1979). “Behavioral Implications of Mechanistic Ecology II: The African Rainbow Lizard, Agama agama.” Copeia, 1979(4), 594--619. ISSN 00458511, 19385110, http://www.jstor.org/stable/1443867.

Riddell EA, Apanovitch EK, Odom JP, Sears MW (2017). “Physical calculations of resistance to water loss improve predictions of species range models.” Ecological Monographs, 87(1), 21-33. doi: 10.1002/ecm.1240 , https://esajournals.onlinelibrary.wiley.com/doi/pdf/10.1002/ecm.1240, https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1002/ecm.1240.

Roughgarden J (1981). “Resource partitioning of space and its relationship to body temperature in Anolis lizard populations.” Oecologia, 50, 256 -- 264. https://link.springer.com/article/10.1007/BF00348048.

Whitford WG, Hutchison VH (1967). “Body Size and Metabolic Rate in Salamanders.” Physiological Zoology, 40(2), 127-133. doi: 10.1086/physzool.40.2.30152447 , https://doi.org/10.1086/physzool.40.2.30152447.

See also

Examples

  surface_area_from_mass(M = 1:50, taxon = "lizard")
  surface_area_from_mass(M = 1:50, taxon = "salamander")
  surface_area_from_mass(M = 1:50, taxon = "frog")
  surface_area_from_mass(M = seq(0.1, 5, 0.1), taxon = "insect")