Organism Volume from Length

Estimate volume (m3) from length (m) for a variety of taxa following Mitchell (1976) .

volume_from_length(l, taxon)

Arguments

l
numeric length (m).
Snout-vent length for amphibians and reptiles is used, except for turtles (carapace length is used).

taxon
character taxon of organism, current choices: "lizard", "frog", "sphere".

Value

numeric volume (m3).

Details

Relationships come from

Lizards: Norris (1965)

Frogs: Tracy (1972)

Sphere: Mitchell (1976)

References

Mitchell JW (1976). “Heat transfer from spheres and other animal forms.” Biophysical Journal, 16(6), 561--569. ISSN 0006-3495, doi: 10.1016/S0006-3495(76)85711-6 , https://www.sciencedirect.com/science/article/pii/S0006349576857116.

Norris KS (1965). “Color adaptation in desert reptiles and its thermal relationships.” In Symposium on Lizard Ecology, 162--229. U. Missouri Press.

Tracy CR (1972). “Newton's Law: Its Application for Expressing Heat Losses from Homeotherms.” BioScience, 22(11), 656--659. ISSN 0006-3568, doi: 10.2307/1296267 , https://academic.oup.com/bioscience/article-pdf/22/11/656/626571/22-11-656.pdf.

See also

Examples

  volume_from_length(l     = 0.05,  
                     taxon = "lizard")

#> [1] 3.478309e-06

  volume_from_length(l     = 0.05,   
                     taxon = "frog")

#> [1] 1.068642e-05

  volume_from_length(l     = 0.05,   
                     taxon = "sphere")

#> [1] 6.556091e-05