Wind Speed at a Specific Height Under Neutral Conditions

Calculate wind speed (m s-1) at a specified height (m) within a boundary layer near the surface. The profile assumes neutral conditions. The velocity profile is the neutral profile described by Sellers (1965) . Function is equations (2) and (3) of Porter et al. (1973) .

wind_speed_profile_neutral(u_r, zr, z0, z)

Arguments

u_r
numeric wind velocity (m s-1) at reference height.

zr
numeric initial reference height (m).

z0
numeric surface roughness (m).

z
numeric height to scale (m).

Value

numeric windspeed (m s-1).

References

Porter WP, Mitchell JW, Bekman A, DeWitt CB (1973). “Behavioral implications of mechanistic ecology: thermal and behavioral modeling of desert ectotherms and their microenvironments.” Oecologia, 13, 1--54.

Sellers WD (1965). Physical climatology. University of Chicago Press, Chicago, IL, USA.

See also

Examples

  wind_speed_profile_neutral(u_r = 0.1, 
                             zr  = 0.1, 
                             z0  = 0.2, 
                             z   = 0.15)

#> [1] 0.1380182