DRYAIR

Calculates several properties of dry air and related characteristics shownas output variables below. The program is based on equations from List, R. J. 1971. Smithsonian Meteorological Tables. Smithsonian Institution Press. Washington, DC. WETAIR must be used in conjunction with function VAPPRS.

The user must supply values for the input variables (db, bp and alt). If alt is known (-1000 < alt < 20000) but not BP, then set BP=0

DRYAIR(db = db, bp = 0, alt = 0)

Arguments

db

Dry bulb temperature (C)

bp

Barometric pressure (pascal)

alt

Altitude (m)

Value

patmos Standard atmospheric pressure (Pa) densty Density (kg m-3) visdyn Dynamic viscosity (kg m-1 s-1) viskin Kinematic viscosity (m2 s-1) difvpr Diffusivity of water vapor in air (m2 s-1) thcond Thermal conductivity (W m-1 K-1) htovpr Latent heat of vapourisation of water (J kg-1) tcoeff Temperature coefficient of volume expansion (K-1) ggroup Group of variables in Grashof number (1-m3 -K) bbemit black body emittance (W m-2) emtmax Wave length of maximum emittance (m)

Examples

  DRYAIR(db=30, bp=100*1000, alt=0)
#> $patmos
#> [1] 101325
#> 
#> $densty
#> [1] 1.149212
#> 
#> $visdyn
#> [1] 1.866409e-05
#> 
#> $viskin
#> [1] 1.624077e-05
#> 
#> $difvpr
#> [1] 2.729119e-05
#> 
#> $thcond
#> [1] 0.0263614
#> 
#> $htovpr
#> [1] 2429839
#> 
#> $tcoeff
#> [1] 0.003298697
#> 
#> $ggroup
#> [1] 1226388
#> 
#> $bbemit
#> [1] 478.8963
#> 
#> $emtmax
#> [1] 9.556325e-06
#>