Properties of Dry Air

The function calculates several properties of dry air and related characteristics shown as output variables below. The program is based on equations from List (1971) and code implementation from NicheMapR (Kearney and Porter 2017; Kearney and Porter 2020) .

The user must supply values for the input variables db, bp, and alt. If alt is known (-1000 < alt < 20000) but not BP, then set bp = 0.

DRYAIR(db, bp = 0, alt = 0)

Arguments

db

numeric dry bulb temperature (C).

bp

numeric barometric pressure (Pa).

alt

numeric altitude (m).

Value

Named list with elements

patmos: numeric standard atmospheric pressure (Pa)

density: numeric density (kg m-3)

visdyn: numeric dynamic viscosity (kg m-1 s-1)

viskin: numeric kinematic viscosity (m2 s-1)

difvpr: numeric diffusivity of water vapor in air (m2 s-1)

thcond: numeric thermal conductivity (W K-1 m-1)

htovpr: numeric latent heat of vaporization of water (J kg-1)

tcoeff: numeric temperature coefficient of volume expansion (K-1)

ggroup: numeric group of variables in Grashof number (m-3 K-1)

bbemit: numeric black body emittance (W m-2)

emtmax: numeric wave length of maximum emittance (m)

References

Kearney MR, Porter WP (2017). “NicheMapR - an R package for biophysical modelling: the microclimate model.” Ecography, 40, 664-674. doi:10.1111/ecog.02360 .

Kearney MR, Porter WP (2020). “NicheMapR - an R package for biophysical modelling: the ectotherm and Dynamic Energy Budget models.” Ecography, 43(1), 85-96. doi:10.1111/ecog.04680 .

List RJ (1971). “Smithsonian Meteorological Tables.” Smithsonian Miscellaneous Collections, 114(1), 1-527. https://repository.si.edu/handle/10088/23746.

Examples

  DRYAIR(db = 30, 
         bp = 100*1000, 
         alt = 0)
#> $patmos
#> [1] 101325
#> 
#> $density
#> [1] 1.149212
#> 
#> $visdyn
#> [1] 1.866409e-05
#> 
#> $viskin
#> [1] 1.624077e-05
#> 
#> $difvpr
#> [1] 2.729119e-05
#> 
#> $thcond
#> [1] 0.0263614
#> 
#> $htovpr
#> [1] 2429839
#> 
#> $tcoeff
#> [1] 0.003298697
#> 
#> $ggroup
#> [1] 1226388
#> 
#> $bbemit
#> [1] 478.8963
#> 
#> $emtmax
#> [1] 9.556325e-06
#>