Construct a Gaussian-Quadratic Thermal Performance Curve.

Constructs a thermal performance curve by combining as a gaussian function to describe the rise in performance up to the optimal temperature and a quadratic decline to zero performance at critical thermal maxima and higher temperatures. Reference: (CA Deutsch 2008)

TPC(T, Topt, CTmin, CTmax)

Arguments

T

numeric vector of temperature range (C)

Topt

numeric thermal optima (C), the temperature at which peak performance occurs

CTmin

numeric critical thermal minima (C), the lower temperature limit for performance

CTmax

numeric critical thermal maxima (C), the upper temperature limit for performance

Value

performance

Details

Construct a Gaussian-quadratic thermal performance curve.

References

CA Deutsch JTRHKSCGDHPM (2008). “Impacts of climate warming on terrestrial ectotherms across latitude.” Proceedings of the National Academy of Science of the United States of America, 105, 6668–6672. doi: 10.1073/pnas.0709472105 , https://www.pnas.org/content/105/18/6668.

Examples

  TPC(T     = 0:60, 
      Topt  = 30, 
      CTmin = 10, 
      CTmax = 40)
#>  [1] 0.0001234098 0.0002226299 0.0003936690 0.0006823281 0.0011592292
#>  [6] 0.0019304541 0.0031511116 0.0050417603 0.0079070541 0.0121551783
#> [11] 0.0183156389 0.0270518469 0.0391638951 0.0555762126 0.0773047404
#> [16] 0.1053992246 0.1408584209 0.1845195240 0.2369277587 0.2981972794
#> [21] 0.3678794412 0.4448580662 0.5272924240 0.6126263942 0.6976763261
#> [26] 0.7788007831 0.8521437890 0.9139311853 0.9607894392 0.9900498337
#> [31] 1.0000000000 0.9900000000 0.9600000000 0.9100000000 0.8400000000
#> [36] 0.7500000000 0.6400000000 0.5100000000 0.3600000000 0.1900000000
#> [41] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [46] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [51] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [56] 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> [61] 0.0000000000