Predict Body Temperature (Operative Environmental Temperature) of an ectotherm in K.

Predict Body Temperature (Operative Environmental Temperature) of an ectotherm in K.

Tb_Gates(
  A,
  D,
  psa_dir,
  psa_ref,
  psa_air,
  psa_g,
  T_g,
  T_a,
  Qabs,
  epsilon,
  H_L,
  ef = 1.3,
  K
)

Arguments

A

numeric surface area (m^2)

D

numeric characteristic dimension for conduction (meters)

psa_dir

numeric proportion surface area exposed to direct radiation from the sky (or enclosure) (0-1)

psa_ref

numeric proportion surface area exposed to reflected radiation from the ground (0-1)

psa_air

numeric proportion surface area exposed to air (0-1)

psa_g

numeric proportion surface area in contact with substrate (0-1)

T_g

numeric ground surface temperature (K)

T_a

numeric ambient air temperature (K)

Qabs

numeric Solar radiation absorbed (W)

epsilon

numeric longwave infrared emissivity of skin (proportion), 0.95 to 1 for most animals (Gates 1980)

H_L

numeric Convective heat transfer coefficient (W m^-2 K^-1)

ef

numeric enhancement factor used to adjust H_L to field conditions (using h_L approximation from Mitchell 1976). Approximated as 1.23 by default, but see Mitchell 1976 for relationship.

K

numeric Thermal conductivity (W K^-1 m^-1 ), K=0.5 W K^-1 m^-1 for naked skin, K=0.15 for insect cuticle (Galushko et al 2005); conductivity of the ground is generally greater than that of animal tissues, so animal thermal conductivity is generally the rate limiting step.

Value

numeric operative environmental temperature (K)

Details

Predicts body temperatures (operative environmental temperature) of an ectotherm in K. Uses approximation in (Gates 1980) . Omits evaporative and metabolic heat loss. Other sources: (Mitchell 1976) , (Kingsolver 1983)

References

Gates DM (1980). Biophysical Ecology. Springer-Verlag, New York, NY, USA.

Kingsolver JG (1983). “Thermoregulation and Flight in Colias Butterflies: Elevational Patterns and Mechanistic Limitations.” Ecology, 64(3), 534-545. doi: 10.2307/1939973 , https://esajournals.onlinelibrary.wiley.com/doi/pdf/10.2307/1939973, https://esajournals.onlinelibrary.wiley.com/doi/abs/10.2307/1939973.

Mitchell JW (1976). “Heat transfer from spheres and other animal forms.” Biophysical Journal, 16(6), 561-569. ISSN 0006-3495, doi: 10.1016/S0006-3495(76)85711-6 , https://www.sciencedirect.com/science/article/pii/S0006349576857116.

See also

Examples

 Tb_Gates (A       = 1, 
           D       = 0.001, 
           psa_dir = 0.6, 
           psa_ref = 0.4, 
           psa_air = 0.6, 
           psa_g   = 0.2, 
           T_g     = 303, 
           T_a     = 310, 
           Qabs    = 2, 
           epsilon = 0.95, 
           H_L     = 10, 
           ef      = 1.23, 
           K       = 0.5)
#> [1] 303.3096