Operative Environmental Temperature of a Butterfly

Predicts body temperatures (operative environmental temperatures) of a butterfly in C.

Based on Kingsolver (1983) and Buckley and Kingsolver (2012) .

Tb_butterfly(
  T_a,
  Tg,
  Tg_sh,
  u,
  H_sdir,
  H_sdif,
  z,
  D,
  delta,
  alpha,
  r_g = 0.3,
  shade = FALSE
)

Arguments

T_a

numeric air temperature in C.

Tg

numeric surface temperature in C in the sunlight.

Tg_sh

numeric surface temperature in C in the shade.

u

numeric wind speed in m / s.

H_sdir

numeric direct solar radiation flux in W / m2.

H_sdif

numeric diffuse solar radiation flux in W / m2.

z

numeric solar zenith angle in degrees.

D

numeric thoracic diameter in cm.

delta

numeric thoracic fur thickness in mm.

alpha

numeric wing solar absorbtivity as a proportion. Range for Colias is 0.4 to 0.7.

r_g

numeric substrate solar reflectivity (proportion), see Kingsolver (1983) .

shade

logical indicator whether body temperature should be calculate in sun (FALSE) or shade (TRUE).

Value

numeric predicted body (operative environmental) temperature (C).

Details

Thermal radiative flux is calculated following Gates (1980) based on Swinbank (1960) and Kingsolver (1983) estimates using the Brunt equation with black body sky temperature from Swinbank (1963) .

References

Buckley LB, Kingsolver JG (2012). “The demographic impacts of shifts in climate means and extremes on alpine butterflies.” Functional Ecology, 26(4), 969-977. doi: 10.1111/j.1365-2435.2012.01969.x , https://besjournals.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1365-2435.2012.01969.x, https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-2435.2012.01969.x.

Gates DM (1980). Biophysical Ecology. Springer-Verlag, New York, NY, USA.

Kingsolver JG (1983). “Thermoregulation and Flight in Colias Butterflies: Elevational Patterns and Mechanistic Limitations.” Ecology, 64(3), 534-545. doi: 10.2307/1939973 , https://esajournals.onlinelibrary.wiley.com/doi/pdf/10.2307/1939973, https://esajournals.onlinelibrary.wiley.com/doi/abs/10.2307/1939973.

Swinbank WC (1960). “Wind profile in thermally stratified flow.” Nature, 186, 463-464.

Swinbank WC (1963). “Long-wave radiation from clear skies.” Quarterly Journal of the Royal Meteorological Society, 89, 339-348.

See also

Examples

  Tb_butterfly(
    T_a = 25, 
    Tg = 25, 
    Tg_sh = 20, 
    u = 0.4, 
    H_sdir = 300, 
    H_sdif = 100, 
    z = 30, 
    D = 0.36, 
    delta = 1.46, 
    alpha = 0.6, 
    r_g = 0.3)
#> [1] 29.16614