Helmuth Lab Calculations for Operative Environmental Temperature of a Limpet

Predicts body temperatures (operative environmental temperatures) of a limpet in C. Calculations from Helmuth Lab -- although radiation and convection are altered from original model -- based on Denny and Harley (2006) .

Tb_limpetBH(T_a, T_r, L, H, I, u, s_aspect, s_slope, c)

Arguments

T_a
numeric air temperature in C.

T_r
numeric rock surface temperature in C in the sunlight.

L
numeric limpet length (anterior/posterior axis) (m).

H
numeric limpet height (dorsal/ventral axis) (m).

I
numeric solar irradiance W / m2.

u
numeric wind speed (m / s).

s_aspect
numeric solar aspect angle (degree), the angle between the limpet's length dimension and the vector to the Sun. Between 70 and 110 degrees.

s_slope
numeric solar elevation angle (degree), the altitude of the Sun, which is the angle between the horizon and the sun.

c
numeric fraction of the sky covered by cloud.

Value

numeric predicted body (operative environmental) temperature (C).

Details

The original equation uses a finite-difference approach where they divide the rock into series of chunks, and calculate the temperature at each node to derive the conductive heat. For simplification, here it takes the rock temperature as a parameter, and conductive heat is calculated by the product of the area, thermal conductivity of rock and the difference in temperatures of the rock and the body.

Limpets are simulated as cones following and using solar emissiviity values from Campbell and Norman (1998) .

The area of the limpet's shell (m2) is projected in the direction at which sunlight strikes the organism Pennell and Deignan (1989) .

Air conductivity values (W / (m-1 K-1)) are calculated following Denny and Harley (2006) .

References

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

Denny MW, Harley CDG (2006). “Hot limpets: predicting body temperature in a conductance-mediated thermal system.” Journal of Experimental Biology, 209(13), 2409--2419. ISSN 0022-0949, doi: 10.1242/jeb.02356 , https://journals.biologists.com/jeb/article-pdf/209/13/i/1255483/isi-iii.pdf.

Pennell S, Deignan J (1989). “Computing the Projected Area of a Cone.” SIAM Review, 31, 299--302. https://epubs.siam.org/doi/10.1137/1031052.

See also

Author

Brian Helmuth lab

Examples

  Tb_limpetBH(T_a = 25, 
              T_r = 30, 
              L = 0.0176, 
              H = 0.0122, 
              I = 1300, 
              u = 1, 
              s_aspect = 90, 
              s_slope = 60, 
              c = 1)

#> [1] 30.68141