Predict Body Temperature (Operative Environmental Temperature) of a Lizard in C

Predicts body temperature (operative environmental temperature) of a lizard in C based on (Campbell and Norman 1998) . Designed for Sceloporus lizards and described in (Buckley 2008) .

Tb_lizard(
  T_a,
  T_g,
  u,
  svl,
  m,
  psi,
  rho_S,
  elev,
  doy,
  sun = TRUE,
  surface = TRUE,
  alpha_S = 0.9,
  alpha_L = 0.965,
  epsilon_s = 0.965,
  F_d = 0.8,
  F_r = 0.5,
  F_a = 0.5,
  F_g = 0.5
)

Arguments

T_a

numeric air temperature (C)

T_g

numeric surface temperature (C)

u

numeric wind speed (m/s)

svl

numeric lizard snout vent length (mm)

m

numeric lizard mass in g, note that it can be estimated as massfromsvl=function(svl) 3.55*10^-5*(svl)^3.00 (Tinkle and Ballinger 1972)

psi

numeric solar zenith angle (degrees)

rho_S

numeric surface albedo (proportion). ~0.25 for grass, ~0.1 for dark soil, >0.75 for fresh snow (Campbell and Norman 1998)

elev

numeric elevation (m)

doy

numeric day of year (1-366)

sun

boolean indicates whether lizard is in sun (TRUE) or shade (FALSE)

surface

boolean indicates whether lizard is on ground surface (TRUE) or above surface (FALSE, e.g. in a tree)

alpha_S

numeric lizard solar absorptivity, alpha_S=0.9 (Gates 1980) (Table 11.4)

alpha_L

numeric lizard thermal absoptivity, alpha_L=0.965 (Bartlett and Gates 1967)

epsilon_s

numeric surface emisivity of lizards, epsilon_s=0.965 (Bartlett and Gates 1967)

F_d

numeric the view factor between the surface of the lizard and diffuse solar radiation (Bartlett and Gates 1967)

F_r

numeric the view factor between the surface of the lizard and reflected solar radiation

F_a

numeric the view factor between the surface of the lizard and atmospheric radiation

F_g

numeric the view factor between the surface of the lizard and ground thermal radation

Value

T_e numeric Operative temperature (C)

References

Bartlett PN, Gates DM (1967). “The energy budget of a lizard on a tree trunk.” Ecology, 48, 316-322.

Buckley LB (2008). “Linking traits to energetics and population dynamics to predict lizard ranges in changing environments.” American Naturalist, 171(1), E1 - E19. doi: 10.1086/523949 , https://pubmed.ncbi.nlm.nih.gov/18171140/.

Campbell GS, Norman JM (1998). Introduction to environmental biophysics, 2nd ed. edition. Springer, New York. ISBN 0387949372.

Gates DM (1980). Biophysical Ecology. Springer-Verlag, New York, NY, USA.

Tinkle DW, Ballinger RE (1972). “Sceloporus Undulatus: A Study of the Intraspecific Comparative Demography of a Lizard.” Ecology, 53(4), 570-584. doi: 10.2307/1934772 , https://esajournals.onlinelibrary.wiley.com/doi/10.2307/1934772.

See also

Examples

  Tb_lizard(T_a       = 25, 
            T_g       = 30, 
            u         = 0.1, 
            svl       = 60, 
            m         = 10, 
            psi       = 34, 
            rho_S     = 0.24, 
            elev      = 500, 
            doy       = 200, 
            sun       = TRUE, 
            surface   = TRUE, 
            alpha_S   = 0.9, 
            alpha_L   = 0.965, 
            epsilon_s = 0.965, 
            F_d       = 0.8, 
            F_r       = 0.5, 
            F_a       = 0.5, 
            F_g       = 0.5)
#> [1] 48.97929