Predicts body temperatures (operative environmental temperature) of a lizard in C.

Predicts body temperature (operative environmental temperature) of a lizard in C. Based on Campbell and Norman (1998, An introduction to environmental biophysics). Designed for Sceloporus lizards and described in Buckley (2008, Linking traits to energetics and population dynamics to predict lizard ranges in changing environments. American Naturalist 171: E1-E19).

Tb_lizard(
  T_a,
  T_g,
  u,
  svl,
  m,
  psi,
  rho_S,
  elev,
  doy,
  sun = TRUE,
  surface = TRUE,
  alpha_S = 0.9,
  alpha_L = 0.965,
  epsilon_s = 0.965,
  F_d = 0.8,
  F_r = 0.5,
  F_a = 0.5,
  F_g = 0.5
)

Arguments

T_a

is air temperature in C

T_g

is surface temperature in C

u

is wind speed in m/s

svl

is lizard snout vent length in mm

m

is lizard mass in g, note that it can be estimated as massfromsvl=function(svl) 3.55*10^-5*(svl)^3.00 (Tinkle and Ballinger 1972)

psi

is solar zenith angle in degrees

rho_S

is surface albedo (proportion). ~0.25 for grass, ~0.1 for dark soil, >0.75 for fresh snow (Campbell & Norman 1998)

elev

is elevation in m

doy

is day of year (1-366)

sun

indicates whether lizard is in sun (TRUE) or shade (FALSE)

surface

indicates whether lizard is on ground surface (TRUE) or above surface (FALSE, e.g. in a tree)

alpha_S

is lizard solar absorptivity, alpha_S=0.9 (Gates 1980, Table 11.4)

alpha_L

is lizard thermal absoptivity, alpha_L=0.965 (Bartlett & Gates 1967)

epsilon_s

is surface emisivity of lizards, epsilon_s=0.965 (Bartlett & Gates 1967)

F_d

is the view factor between the surface of the lizard and diffuse solar radiation (Bartlett & Gates 1967)

F_r

is the view factor between the surface of the lizard and reflected solar radiation

F_a

is the view factor between the surface of the lizard and atmospheric radiation

F_g

is the view factor between the surface of the lizard and ground thermal radation

Value

T_e Operative temperature (C)

Details

Predicts body temperatures (operative environmental temperature) of a lizard in C.

See also

Examples

  Tb_lizard(
  T_a=25, 
  T_g=30, 
  u=0.1, 
  svl=60, 
  m=10, 
  psi=34, 
  rho_S=0.24, 
  elev=500, 
  doy=200, 
  sun=TRUE, 
  surface=TRUE, 
  alpha_S=0.9, 
  alpha_L=0.965, 
  epsilon_s=0.965, 
  F_d=0.8, 
  F_r=0.5, 
  F_a=0.5, 
  F_g=0.5)