Operative Environmental Temperature of a Mussel

The function estimates body temperature (C, operative environmental temperature) of a mussel. The function implements a steady-state model, which assumes unchanging environmental conditions.

Tb_mussel(l, h, T_a, T_g, S, k_d, u, psi, cl, evap = FALSE, group = "solitary")

Arguments

l

numeric mussel length (anterior/posterior axis, m).

h

numeric mussel height (dorsal/ventral axis, m). It is reasonable to assume h = 0.5 * l.

T_a

numeric air temperature (C).

T_g

numeric ground temperature (C).

S

numeric direct solar flux density (W m-2).

k_d

numeric diffuse fraction, proportion of solar radiation that is diffuse.

u

numeric wind speed (m s-1).

psi

numeric solar zenith angle (degrees): can be calculated from zenith_angle.

cl

numeric fraction of the sky covered by cloud.

evap

logical Whether mussel is gaping to evaporatively cool. If TRUE, the function assumes a constant mass loss rate of 5 percent of the initial body mass per hour.

group

character; options are "aggregated": mussels living in beds; "solitary": solitary individual, anterior or posterior end facing upwind; and "solitary_valve": solitary individual, valve facing upwind.

Value

numeric predicted body (operative environmental) temperature (C).

Details

Thermal radiative flux is calculated following Helmuth (1998) , Helmuth (1999) , and Idso and Jackson (1969) .

References

Helmuth B (1999). “Thermal biology of rocky intertidal mussels: quantifying body temperatures using climatological data.” Ecology, 80(1), 15-34. https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1890/0012-9658%281999%29080%5B0015%3ATBORIM%5D2.0.CO%3B2.

Helmuth BST (1998). “Intertidal Mussel Microclimates: Predicting the Body Temperature of a Sessile Invertebrate.” Ecological Monographs, 68(1), 51--74. ISSN 00129615, doi:10.2307/2657143 .

Idso SB, Jackson RD (1969). “Thermal radiation from the atmosphere.” Journal of Geophysical Research (1896-1977), 74(23), 5397-5403. doi:10.1029/JC074i023p05397 .

See also

Examples

Tb_mussel(l     = 0.1, 
          h     = 0.05, 
          T_a   = 25, 
          T_g   = 30, 
          S     = 500, 
          k_d   = 0.2, 
          u     = 2, 
          psi   = 30, 
          evap  = FALSE, 
          cl    = 0.5, 
          group = "solitary")
#> [1] 34.97406