Predicts body temperature (operative environmental temperature) of a marine snail in °C.

Predicts body temperature of a snail in °C. Implements a steady‐state model, which assumes unchanging environmental conditions. Based on Iacarella and Helmuth 2012. Body temperature and desiccation constrain the activity of Littoraria irrorata within the Spartina alterniflora canopy. Function was provided by the Helmuth lab and is a simplified version of model in publication.

Tb_snail(temp, Len, solar, WS, CC, WL, WSH)

Arguments

temp
air temperature (°C)

Len
snail length (m)

solar
direct solar flux density (W/m2)

WS
wind speed (m/s)

CC
fraction of the sky covered by cloud (0-1)

WL
water loss rate (kg/s), 5 percent loss of body mass over one hour is a reasonable maximum level (Helmuth 1999)

WSH
wind sensor height (m)

Value

predicted body temperature (°C)

Details

Predicts body temperature of a snail in °C.

See also

Author

Brian Helmuth Lab

Examples

if (FALSE) {
Tb_snail(temp = 25, Len = 0.012, solar=800, WS=1, CC=0.5, WL=0, WSH=10)
}