WETAIR

Calculates several properties of humid air as output variables below. The program is based on equations from List, R. J. 1971. Smithsonian Meteorological Tables. Smithsonian Institution Press. Washington, DC. WETAIR must be used in conjunction with function VAPPRS.

Input variables are shown below. The user must supply known values for DB and BP (BP at one standard atmosphere is 101,325 pascals). Values for the remaining variables are determined by whether the user has either (1) psychrometric data (WB or RH), or (2) hygrometric data (DP)

(1) Psychrometric data: If WB is known but not RH, then set RH = -1 and DP = 999. If RH is known but not WB then set WB = 0 and DP = 999. (2) Hygrometric data: If DP is known then set WB = 0 and RH = 0.

WETAIR(db = db, wb = db, rh = 0, dp = 999, bp = 101325)

Arguments

db

Dry bulb temperature (C)

wb

Wet bulb temperature (C)

rh

Relative humidity (%)

dp

Dew point temperature (C)

bp

Barometric pressure (pascal)

Value

e Vapour pressure (Pa) esat Saturation vapour pressure (Pa) vd Vapour density (kg m-3) rw Mixing ration (kg kg-1) tvir Virtual temperature (K) tvinc Virtual temperature increment (K) denair Hourly predictions of the soil moisture under the maximum specified shade cp Specific heat of air at constant pressure (J kg-1 K-1) wtrpot Water potential (Pa) Relative humidity (%)

Examples

WETAIR(db=30, wb=28, rh=60, bp=100*1000)
#> $e
#> [1] 2545.82
#> 
#> $esat
#> [1] 4243.033
#> 
#> $vd
#> [1] 0.01823351
#> 
#> $rw
#> [1] 0.01633625
#> 
#> $tvinc
#> [1] 2.961613
#> 
#> $denair
#> [1] 1.139221
#> 
#> $cp
#> [1] 1018.367
#> 
#> $wtrpot
#> [1] -71466408
#> 
#> $rh
#> [1] 60
#>