Properties of Wet Air

The function calculates several properties of humid air described as output variables below. The program is based on equations from List (1971) and code implementation from NicheMapR (Kearney and Porter 2017; Kearney and Porter 2020) .
WETAIR must be used in conjunction with VAPPRS. Input variables are shown below. See Details.

WETAIR(db, wb = db, rh = 0, dp = 999, bp = 101325)

Arguments

db

numeric dry bulb temperature (C).

wb

numeric wet bulb temperature (C).

rh

numeric relative humidity (%).

dp

numeric dew point temperature (C).

bp

numeric barometric pressure (Pa).

Value

Named list with elements

e: numeric saturation vapor pressure (Pa)

vd: numeric vapor density (kg m-3)

rw: numeric mixing ratio (kg kg-1)

tvir: numeric virtual temperature (K)

tvinc: numeric virtual temperature increment (K)

denair: numeric density of the air (kg m-3)

cp: numeric specific heat of air at constant pressure (J kg-1 K-1)

wtrpot: numeric water potential (Pa)

rh: numeric relative humidity (%)

Details

The user must supply known values for DB and BP (BP at one standard atmosphere is 101,325 pascals). Values for the remaining variables are determined by whether the user has either (1) psychrometric data (WB or RH), or (2) hygrometric data (DP):

Psychrometric data: If WB is known but not RH, then set RH = -1 and DP = 999. If RH is known but not WB then set WB = 0 and DP = 999

Hygrometric data: If DP is known, set WB = 0 and RH = 0

References

Kearney MR, Porter WP (2017). “NicheMapR - an R package for biophysical modelling: the microclimate model.” Ecography, 40, 664-674. doi:10.1111/ecog.02360 .

Kearney MR, Porter WP (2020). “NicheMapR - an R package for biophysical modelling: the ectotherm and Dynamic Energy Budget models.” Ecography, 43(1), 85-96. doi:10.1111/ecog.04680 .

List RJ (1971). “Smithsonian Meteorological Tables.” Smithsonian Miscellaneous Collections, 114(1), 1-527. https://repository.si.edu/handle/10088/23746.

Examples

  WETAIR(db = 30, 
         wb = 28, 
         rh = 60, 
         bp = 100 * 1000)
#> $e
#> [1] 2544.359
#> 
#> $esat
#> [1] 4240.599
#> 
#> $vd
#> [1] 0.01822304
#> 
#> $rw
#> [1] 0.01632663
#> 
#> $tvinc
#> [1] 2.959897
#> 
#> $denair
#> [1] 1.139227
#> 
#> $cp
#> [1] 1018.359
#> 
#> $wtrpot
#> [1] -71466408
#> 
#> $rh
#> [1] 60
#>