Calculate Properties of Wet Air

Calculates several properties of humid air as output variables below. The program is based on equations from (List 1971) .
WETAIR must be used in conjunction with function VAPPRS. Input variables are shown below. See Details.

WETAIR(db, wb = db, rh = 0, dp = 999, bp = 101325)

Arguments

db
numeric Dry bulb temperature (C).

wb
numeric Wet bulb temperature (C).

rh
numeric Relative humidity (%).

dp
numeric Dew point temperature (C).

bp
numeric Barometric pressure (Pa).

Value

Named list with elements

e: numeric saturation vapor pressure (Pa)

vd: numeric vapor density (kg m-3)

rw: numeric mixing ration (kg kg-1)

tvir: numeric virtual temperature (K)

tvinc: numeric virtual temperature increment (K)

denair: denair numeric hourly predictions of the soil moisture under the maximum specified shade

cp: numeric specific heat of air at constant pressure (J kg-1 K-1)

wtrpot: numeric water potential (Pa)

rh: numeric relative humidity (%)

Details

The user must supply known values for DB and BP (BP at one standard atmosphere is 101,325 pascals). Values for the remaining variables are determined by whether the user has either (1) psychrometric data (WB or RH), or (2) hygrometric data (DP):

Psychrometric data: If WB is known but not RH, then set RH = -1 and DP = 999. If RH is known but not WB then set WB = 0 and DP = 999

Hygrometric data: If DP is known, set WB = 0 and RH = 0

References

List RJ (1971). “Smithsonian Meteorological Tables.” Smithsonian Miscellaneous Collections, 114(1), 1--527. https://repository.si.edu/handle/10088/23746.

Examples

  WETAIR(db = 30, 
         wb = 28, 
         rh = 60, 
         bp = 100 * 1000)

#> $e
#> [1] 2545.82
#> 
#> $esat
#> [1] 4243.033
#> 
#> $vd
#> [1] 0.01823351
#> 
#> $rw
#> [1] 0.01633625
#> 
#> $tvinc
#> [1] 2.961613
#> 
#> $denair
#> [1] 1.139221
#> 
#> $cp
#> [1] 1018.367
#> 
#> $wtrpot
#> [1] -71466408
#> 
#> $rh
#> [1] 60
#>