Air Temperature at a Specified Height

The function calculates air temperature (C) at a specified height (m). Estimates a three segment velocity and temperature profile based on user-specified, experimentally determined values for 3 roughness heights and reference heights. Multiple heights are appropriate in heterogenous areas with, for example, a meadow, bushes, and rocks. Implements the MICROSEGMT routine from NicheMapR as described in Kearney and Porter (2017) .

air_temp_profile_segment(T_r, u_r, zr, z0, z, T_s)

Arguments

T_r

numeric a vector of air temperatures (C) at the 3 reference heights.

u_r

numeric a vector of wind speeds (m s-1) at the 3 reference heights.

zr

numeric a vector of 3 reference heights (meters).

z0

numeric a vector of 3 experimentally determined roughness heights (meters).

z

numeric height for air temperature estimation (meters).

T_s

numeric surface temperatures (C).

Value

numeric air temperature (C).

References

Kearney MR, Porter WP (2017). “NicheMapR - an R package for biophysical modelling: the microclimate model.” Ecography, 40, 664-674. doi:10.1111/ecog.02360 .

See also

Examples

  air_temp_profile_segment(T_r = c(25, 22, 20), 
                           u_r = c(0.01, 0.025, 0.05), 
                           zr  = c(0.05, 0.25, 0.5), 
                           z0  = c(0.01, 0.15, 0.2), 
                           z   = 0.3, 
                           T_s = 27)
#> [1] 21.40057