Estimate Diurnal Temperatures

Estimate temperature across hours using a diurnal temperature variation function incorporating sine and exponential components (Parton and Logan 1981) . Default alpha, beta, gamma values are the average of 5 North Carolina sites (M. Wann and Gold 1985) . Other alpha, beta, gamma parameterizations include values for Denver, Colorado from (Parton and Logan 1981) : 150cm air temperature: alpha = 1.86, beta = 2.20, gamma = -0.17; 10cm air temperature: alpha = 1.52, beta = 2.00, gamma = -0.18; soil surface temperature: alpha = 0.50, beta = 1.81, gamma = 0.49; 10cm soil temperature: alpha = 0.45, beta = 2.28, gamma = 1.83.

diurnal_temp_variation_sineexp(
  T_max,
  T_min,
  t,
  t_r,
  t_s,
  alpha = 2.59,
  beta = 1.55,
  gamma = 2.2
)

Arguments

T_max, T_min

numeric maximum and minimum daily temperatures (C).

t

numeric time for temperature estimate (hour).

t_r, t_s

numeric times of sunrise and sunset (hour).

alpha

numeric time difference between t_x (time of maximum temperature) and noon (hour).

beta

numeric time difference between t_x and sunrise (hour).

gamma

numeric decay parameter for rate of t change from sunset to t_n (time of minimum temp).

Value

numeric temperature (C) at a specified hour.

References

M. Wann DY, Gold HJ (1985). “Evaluation and calibration of three models for daily cycle of air temperature.” Agricultural and Forest Meteorology, 34, 121-128. https://agris.fao.org/agris-search/search.do?recordID=NL19850065824.

Parton WJ, Logan JA (1981). “A model for diurnal variation in soil and air temperature.” Agricultural Meteorology, 23, 205-216. doi: 10.1016/0002-1571(81)90105-9 , https://www.sciencedirect.com/science/article/abs/pii/0002157181901059.

See also

Examples

  diurnal_temp_variation_sineexp(T_max = 30, 
                                 T_min = 10, 
                                 t     = 11, 
                                 t_r   = 6, 
                                 t_s   = 18, 
                                 alpha = 2.59, 
                                 beta  = 1.55, 
                                 gamma = 2.2)
#> [1] 23.91954