Soil Thermal Conductivity

The function estimates soil thermal conductivity (W m-1 K-1) using the methods of deVries (1963) .

soil_conductivity(x, lambda, g_a)

Arguments

x

numeric vector of volume fractions of soil constituents (e.g., clay, quartz, minerals other than quartz, organic matter, water, air). The volume fractions should sum to 1. Note that x and lambda values in the example correspond to these soil constituents.

lambda

numeric vector of the thermal conductivities (W m-1 K-1) of the soil constituents.

g_a

numeric shape factor on soil particles. The soil particles are assumed to be ellipsoids with axes g_a, g_b, and g_c, where g_a + g_b + g_c = 1 and g_a = g_b. deVries (1952) suggests g_a = g_b = 0.125.

Value

numeric soil thermal conductivity (W m-1 K-1).

References

deVries DA (1952). “Thermal Conductivity of Soil.” Nature, 178, 1074. doi:10.1038/1781074a0 .

deVries DA (1963). “Thermal Properties of Soils.” In Physics of Plant Environment. North Holland Publishing Company. doi:10.1002/qj.49709038628 .

See also

Author

Joseph Grigg

Examples

  soil_conductivity(x      = c(0.10, 0.40, 0.11, 0.01, 0.2, 0.18), 
                    lambda = c(0.10, 0.40, 0.11, 0.01, 0.2, 0.18), 
                    g_a     = 0.125)
#> [1] 0.2336174