Solve Equation for Soil Temperature

This function is called by soil_temperature_equation to solve the equation for soil temperature from Beckman et al. (1973) . The function represents the integrand in the equation. It is not intended to be called directly.

soil_temperature_integrand(x, L, z0)

Arguments

x

numeric vector of volume fractions of soil constituents (e.g., clay, quartz, minerals other than quartz, organic matter, water, air). The volume fractions should sum to 1. Note that x and lambda values in the example correspond to these soil constituents.

L

numeric Monin-Obukhov length, a measure of the instability of heat flow (Beckman et al. 1973) .

z0

numeric surface roughness (m).

Value

numeric integrand for soil temperature function.

References

Beckman WA, Mitchell JW, Porter WP (1973). “Thermal Model for Prediction of a Desert Iguana's Daily and Seasonal Behavior.” Journal of Heat Transfer, 95(2), 257-262. doi:10.1115/1.3450037 .

See also

Author

Joseph Grigg

Examples

  soil_temperature_integrand(x  = c(0.10, 0.40, 0.11, 0.01, 0.2, 0.18), 
                             L  = -10, 
                             z0 = 0.2)
#> [1]   9.470192   7.044393   8.613718 -63.021412   5.467988   5.800205