Allometry: Surface Area from Mass

Estimate surface area (m2) from mass (g) for one of a variety of taxa.

surface_area_from_mass(M, taxon)

Arguments

M
numeric vector of mass (g).

taxon
character taxonomic classification of organism, current choices: "lizard", "salamander", "frog", "insect".

Value

numeric surface area (m2).

Details

All models follow (SA = a Mb) with mass in grams and surface area in meters2.

Lizards (Norris 1965; Porter and James 1979; Roughgarden 1981; OConnor 1999; Fei et al. 2012) :

\(a = 0.000314 \pi\)
\(b = 2/3\)

Salamanders (Whitford and Hutchison 1967; Riddell et al. 2017) :

\(a = 0.000842\)
\(b = 0.694\)

Frogs (McClanahan and Baldwin 1969) :

\(a = 0.00099\)
\(b = 0.56\)

Insects (Lactin et al. 1997) .

\(a = 0.0013\)
\(b = 0.8\)

References

Fei T, Skidmore AK, Venus V, Wang T, Schlerf M, Toxopeus B, van Overjijk S, Bian M, Liu Y (2012). “A body temperature model for lizards as estimated from the thermal environment.” Journal of Thermal Biology, 37(1), 56--64. ISSN 0306-4565, doi: 10.1016/j.jtherbio.2011.10.013 , https://www.sciencedirect.com/science/article/pii/S0306456511001513.

Lactin DJ, L. D, Johnson (1997). “Response of body temperature to solar radiation in restrained nymphal migratory grasshoppers (Orthoptera: Acrididae): influences of orientation and body size.” Physiological Entomology, 22(2), 131--139. doi: 10.1111/j.1365-3032.1997.tb01150.x , https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1365-3032.1997.tb01150.x.

McClanahan L, Baldwin R (1969). “Rate of water uptake through the integument of the desert toad, Bufo punctatus.” Comparative Biochemistry and Physiology, 28(1), 381--389. ISSN 0010-406X, doi: 10.1016/0010-406X(69)91351-6 , https://www.sciencedirect.com/science/article/pii/0010406X69913516.

Norris KS (1965). “Color adaptation in desert reptiles and its thermal relationships.” In Symposium on Lizard Ecology, 162--229. U. Missouri Press.

OConnor M (1999). “Physiological and ecological implications of a simple model of heating and cooling in reptiles.” Journal of Thermal Biology, 24, 113--136.

Porter WP, James FC (1979). “Behavioral Implications of Mechanistic Ecology II: The African Rainbow Lizard, Agama agama.” Copeia, 1979(4), 594--619. ISSN 00458511, 19385110, http://www.jstor.org/stable/1443867.

Riddell EA, Apanovitch EK, Odom JP, Sears MW (2017). “Physical calculations of resistance to water loss improve predictions of species range models.” Ecological Monographs, 87(1), 21--33. doi: 10.1002/ecm.1240 , https://esajournals.onlinelibrary.wiley.com/doi/abs/10.1002/ecm.1240.

Roughgarden J (1981). “Resource partitioning of space and its relationship to body temperature in Anolis lizard populations.” Oecologia, 50, 256--264. https://link.springer.com/article/10.1007/BF00348048.

Whitford WG, Hutchison VH (1967). “Body Size and Metabolic Rate in Salamanders.” Physiological Zoology, 40(2), 127--133. doi: 10.1086/physzool.40.2.30152447 , https://doi.org/10.1086/physzool.40.2.30152447.

See also

Examples

  surface_area_from_mass(M     = 1:50, 
                         taxon = "lizard")

#>  [1] 0.0009864601 0.0015659078 0.0020519197 0.0024857237 0.0028844268
#>  [6] 0.0032572195 0.0036097591 0.0039458404 0.0042681649 0.0045787422
#> [11] 0.0048791179 0.0051705136 0.0054539157 0.0057301353 0.0059998495
#> [16] 0.0062636312 0.0065219701 0.0067752895 0.0070239582 0.0072683001
#> [21] 0.0075086014 0.0077451169 0.0079780743 0.0082076787 0.0084341152
#> [26] 0.0086575515 0.0088781408 0.0090960228 0.0093113257 0.0095241675
#> [31] 0.0097346569 0.0099428947 0.0101489742 0.0103529822 0.0105549995
#> [36] 0.0107551017 0.0109533594 0.0111498387 0.0113446019 0.0115377072
#> [41] 0.0117292099 0.0119191618 0.0121076119 0.0122946066 0.0124801900
#> [46] 0.0126644036 0.0128472870 0.0130288778 0.0132092119 0.0133883233

  surface_area_from_mass(M = 1:50,  
                         taxon = "salamander")

#>  [1] 0.000842000 0.001362156 0.001804821 0.002203645 0.002572748 0.002919773
#>  [7] 0.003249452 0.003564975 0.003868623 0.004162096 0.004446708 0.004723500
#> [13] 0.004993313 0.005256842 0.005514669 0.005767284 0.006015111 0.006258514
#> [19] 0.006497811 0.006733284 0.006965179 0.007193718 0.007419100 0.007641501
#> [25] 0.007861084 0.008077995 0.008292368 0.008504323 0.008713975 0.008921425
#> [31] 0.009126770 0.009330097 0.009531489 0.009731022 0.009928767 0.010124790
#> [37] 0.010319154 0.010511917 0.010703133 0.010892855 0.011081131 0.011268007
#> [43] 0.011453526 0.011637729 0.011820656 0.012002343 0.012182825 0.012362136
#> [49] 0.012540307 0.012717369

  surface_area_from_mass(M = 1:50,  
                         taxon = "frog")

#>  [1] 0.000990000 0.001459527 0.001831568 0.002151735 0.002438139 0.002700225
#>  [7] 0.002943677 0.003172237 0.003388527 0.003594473 0.003791535 0.003980858
#> [13] 0.004163355 0.004339772 0.004510724 0.004676731 0.004838231 0.004995602
#> [19] 0.005149170 0.005299221 0.005446005 0.005589744 0.005730635 0.005868856
#> [25] 0.006004566 0.006137907 0.006269009 0.006397992 0.006524963 0.006650022
#> [31] 0.006773260 0.006894760 0.007014601 0.007132855 0.007249588 0.007364862
#> [37] 0.007478735 0.007591262 0.007702494 0.007812478 0.007921258 0.008028877
#> [43] 0.008135374 0.008240787 0.008345151 0.008448499 0.008550864 0.008652274
#> [49] 0.008752760 0.008852346

  surface_area_from_mass(M = seq(0.1, 5, 0.1),  
                         taxon = "insect")

#>  [1] 0.0002060361 0.0003587297 0.0004961813 0.0006245847 0.0007466539
#>  [6] 0.0008639017 0.0009772862 0.0010874651 0.0011949160 0.0013000000
#> [11] 0.0014029994 0.0015041403 0.0016036073 0.0017015542 0.0017981104
#> [16] 0.0018933868 0.0019874787 0.0020804695 0.0021724322 0.0022634315
#> [21] 0.0023535248 0.0024427639 0.0025311951 0.0026188604 0.0027057979
#> [26] 0.0027920425 0.0028776261 0.0029625779 0.0030469249 0.0031306921
#> [31] 0.0032139026 0.0032965778 0.0033787379 0.0034604015 0.0035415860
#> [36] 0.0036223078 0.0037025824 0.0037824242 0.0038618468 0.0039408631
#> [41] 0.0040194852 0.0040977248 0.0041755926 0.0042530990 0.0043302539
#> [46] 0.0044070666 0.0044835461 0.0045597008 0.0046355388 0.0047110678